I added ameliorations to posts service

+++ src/back/postsservice/src/main/java/com/linktech/postsservice/postsservice/controllers/CommentsController.java
@@ -3,7 +3,9 @@ package com.linktech.postsservice.postsservice.controllers;
 import java.util.*;
 import java.util.NoSuchElementException;
 
+import com.linktech.postsservice.postsservice.ApiClients.UserClient;
 import com.linktech.postsservice.postsservice.models.Comment;
+import com.linktech.postsservice.postsservice.models.ExternalModels.UserModel;
 import com.linktech.postsservice.postsservice.repositories.*;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +13,13 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import feign.Feign;
+import feign.gson.GsonDecoder;
+import feign.gson.GsonEncoder;
+import feign.okhttp.OkHttpClient;
+import feign.slf4j.Slf4jLogger;
+
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -24,22 +33,28 @@ public class CommentsController {
     @Autowired
     private ICommentsRepository commentsRepository;
 
+    UserClient userClient = Feign.builder()
+    .client(new OkHttpClient())
+    .encoder(new GsonEncoder())
+    .decoder(new GsonDecoder())
+    .logger(new Slf4jLogger(UserClient.class))
+    .target(UserClient.class, "http://localhost:2202/users");
+    
     @PostMapping()
     public Comment createPost(@RequestBody Comment post) {
         return commentsRepository.save(post);
     }
-    
+     
     //Example of request: http://localhost:2303/comments/getPostComments/{postId}
     @GetMapping("/getPostComments/{postId}")
     public ResponseEntity<List<Comment>> gePostComments(@PathVariable("postId") String postId){
        
         try  {
-
             List<Comment> comments = commentsRepository.findByPostId(postId);
             return new ResponseEntity<>(comments, HttpStatus.OK);
 
        }catch (NoSuchElementException e){
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }
     }
 
@@ -53,17 +68,40 @@ public class CommentsController {
         return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }
     }
-    
-    @PutMapping(value="/{id}")
-    public Comment updatePost(@PathVariable String id, @RequestBody Comment comment) {
+
+    @PutMapping(value="/{currentUserId}/{id}")
+    public ResponseEntity<Comment> updatePost(@PathVariable("currentUserId") String currentUserId, @PathVariable("id") String id, @RequestBody Comment comment) {
+        Comment thiscomment = commentsRepository.findById(id).get();
+        if (thiscomment.getUserId() == currentUserId)
+            comment.setId(id);
+        else
+        {
+            UserModel userModel = userClient.get(currentUserId);
+            if (userModel.isAdmin)
+                comment.setId(id);
+            else 
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         
-        comment.setId(id);
-        return commentsRepository.save(comment);
+        return new ResponseEntity<Comment>(commentsRepository.save(comment), HttpStatus.OK);
     }
     
-    @DeleteMapping("/{id}")
-    public void deleteComment(@PathVariable("id") String id)
+    @DeleteMapping("/{currentUserId}/{id}")
+    public ResponseEntity<?> deleteComment(@PathVariable("currentUserId") String currentUserId, 
+        @PathVariable("id") String id)
     {
-        commentsRepository.deleteById(id);
+        Comment thiscomment = commentsRepository.findById(id).get();
+        if (thiscomment.getUserId() == currentUserId)
+            commentsRepository.deleteById(id);
+        else
+        {
+            UserModel userModel = userClient.get(currentUserId);
+            if (userModel.isAdmin)
+                commentsRepository.deleteById(id);
+            else 
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
+        return new ResponseEntity<Comment>(HttpStatus.OK);
     }
 }
